are you sure dialogs implemented add task implemented task card and bucket card improved

+++ src/app/buckets/bucket-card/bucket-card.component.ts
@@ -1,7 +1,9 @@
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
 import { Observable, Subscription } from 'rxjs';
+import { SureDialogComponentComponent } from 'src/app/ui/sure-dialog-component/sure-dialog-component.component';
 import * as GlobalVariables from '../../globals';
 
 @Component({
@@ -15,8 +17,10 @@ export class BucketCardComponent implements OnInit {
   // @Input() description = '';
   // @Input() maxTaskCount = '';
   @Output() delete = new EventEmitter<number>();
-  showBucketOptions = false;
+
   showEditBucket = false;
+  showAddTask = false;
+
   @Input() bucket = {
     bucketId: -1,
     owner: '',
@@ -29,7 +33,11 @@ export class BucketCardComponent implements OnInit {
   subscription = new Subscription();
   path = GlobalVariables.GlobalServerPath;
   headers = new HttpHeaders();
-  constructor(private router: Router, private http: HttpClient) {
+  constructor(
+    private router: Router,
+    private http: HttpClient,
+    private dialog: MatDialog,
+  ) {
     this.headers = this.headers.append('Content-Type', 'application/json');
     this.headers = this.headers.append('Accept', 'application/json');
   }
@@ -37,41 +45,51 @@ export class BucketCardComponent implements OnInit {
     alert(err.message);
     console.log(err);
   };
-  bucketClicked() {
-    this.showBucketOptions = true;
-  }
   onShowTasksClick() {
     this.router.navigateByUrl('tasks/bucket/' + this.bucket.bucketId);
   }
   onEditBucketClick() {
-    this.showBucketOptions = false;
     this.showEditBucket = true;
   }
   onEditBucketClose() {
     this.showEditBucket = false;
-  }
-  onBucketOptionsClose() {
-    this.showBucketOptions = false;
   }
   onBucketEditted(bucket: any) {
     console.log(bucket);
     this.bucket = bucket;
     this.showEditBucket = false;
   }
-  onAddTaskClick() {}
+  onAddTaskClick() {
+    this.showAddTask = true;
+  }
+  onAddTaskClose() {
+    this.showAddTask = false;
+  }
   onDeleteBucketClicked() {
-    console.log(this.bucket.bucketId);
-    let deletePath =
-      this.path + GlobalVariables.BucketsPath + this.bucket.bucketId;
-    this.response$ = this.http.delete(deletePath, { headers: this.headers });
-    this.subscription = this.response$.subscribe(
-      (res) => {
-        console.log(res);
-        this.delete.emit(this.bucket.bucketId);
-        this.showBucketOptions = false;
+    const dialogAnchor = this.dialog.open(SureDialogComponentComponent, {
+      data: {
+        title: 'You are deleting bucket',
+        message: 'Are you sure you want to delete this bucket?',
       },
-      (err) => console.log(err),
-    );
+      // KUPA
+
+      // KUPA
+
+      // KUPA
+    });
+    dialogAnchor.afterClosed().subscribe((e) => {
+      if (!e) return;
+      let deletePath =
+        this.path + GlobalVariables.BucketsPath + this.bucket.bucketId;
+      this.response$ = this.http.delete(deletePath, { headers: this.headers });
+      this.subscription = this.response$.subscribe(
+        (res) => {
+          console.log(res);
+          this.delete.emit(this.bucket.bucketId);
+        },
+        (err) => console.log(err),
+      );
+    });
   }
 
   ngOnInit(): void {}
